publisher interface abstraction and compiling dummy benchmark
