+++ realtime_tools/benchmark/realtime_publishers_benchmark.cpp
@@ -1,0 +1,39 @@
+// Copyright (c) 2025, Brian Jin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Author: Brian Jin
+
+#include <benchmark/benchmark.h>
+
+#include <string>
+
+#include <realtime_tools/realtime_publisher.hpp>
+#include <realtime_tools/wait_free_realtime_publisher.hpp>
+
+static void BM_StringCreation(benchmark::State & state)
+{
+  for (auto _ : state) std::string empty_string;
+}
+// Register the function as a benchmark
+BENCHMARK(BM_StringCreation);
+
+// Define another benchmark
+static void BM_StringCopy(benchmark::State & state)
+{
+  std::string x = "hello";
+  for (auto _ : state) std::string copy(x);
+}
+BENCHMARK(BM_StringCopy);
+
+BENCHMARK_MAIN();
